refactor: consolidate the nuvs components
Every module under `components/Nuvs` had exactly one consumer. Several were
single-branch leaves split out when styling lived in styled-components and a
file per styled component was the cost of doing business; with Tailwind there
is nothing holding them apart.

- `NuvsBlast` absorbs `NuvsBlastError`, `NuvsBlastPending` and
  `NuvsBlastResults`. The three were branches of one state machine over one
  request, so the whole panel now reads top to bottom in one file.
- `NuvsList` absorbs `NuvsItem`, `NuvsOrf` absorbs `NuvsOrfLabel`, and
  `NuvsExport` absorbs `NuvsExportPreview`.

Both stylesheets are gone, and with them two duplicated magic numbers:

- `NuvsItem.css` fixed a row at 75px while `NuvsList` separately told the
  virtualizer to estimate 75. Changing one would have detached the list from
  its own scroll geometry. Now one `ROW_HEIGHT` feeds the estimate, and the
  row fills the slot the virtualizer gives it.
- `NuvsOrf.css` padded the ORF label 15px to line it up with the chart's d3
  offset, which was written as a bare 15 and a bare 30 in the drawing code.
  One `INSET` now feeds all three.

`NuvsExportPreview`'s two inline `style` objects become Tailwind classes —
the last of the styled-components residue in the directory.

`NuvsValues` stays its own module: it is the one thing here with two callers.
`NuvsToolbar`, `NuvsSequence` and `NuvsDetail` stay too — each is a coherent
region rather than a fragment.

The BLAST tests moved off the extracted leaf and onto `NuvsBlast` itself, so
they now cover which state the panel picks as well as how each one renders —
absent, pending, failed, returned, and returned-empty.
