--- apps/web/src/analyses/components/Nuvs/NuvsItem.css
@@ -1,3 +1,0 @@
-.nuvs-item {
-	height: 75px;
-}
