--- apps/web/src/analyses/components/Nuvs/NuvsItem.tsx
@@ -1,45 +1,0 @@
-import NuvsValues from "@analyses/components/Nuvs/NuvsValues";
-import type { FormattedNuvsHit } from "@analyses/types";
-import { cn } from "@app/cn";
-import Badge from "@base/Badge";
-import Box from "@base/Box";
-import "./NuvsItem.css";
-
-type NuVsItemProps = {
-	activeHit?: string;
-	hit: FormattedNuvsHit;
-	setActiveHit: (id: string) => void;
-};
-
-/**
- * A condensed Nuvs item for use in a list of Nuvs
- */
-export default function NuvsItem({
-	activeHit,
-	hit,
-	setActiveHit,
-}: NuVsItemProps) {
-	const { annotatedOrfCount, e, id, index, sequence } = hit;
-
-	const isActive = activeHit !== undefined && Number(activeHit) === id;
-
-	return (
-		<Box
-			className={cn(
-				{ "bg-slate-100": isActive },
-				"border-none",
-				"nuvs-item",
-				"rounded-none",
-				"m-0",
-			)}
-			onClick={() => setActiveHit(String(id))}
-		>
-			<div className="flex items-center justify-between">
-				<span className="font-medium">Sequence {index}</span>
-				<Badge>{sequence.length}</Badge>
-			</div>
-
-			<NuvsValues e={e} orfCount={annotatedOrfCount} />
-		</Box>
-	);
-}
